requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
coverage==5.5
Django==3.2.5
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.10.1
importlib-metadata==4.6.1
inflection==0.5.1
iniconfig==1.1.1
packaging==21.0
Pillow==8.3.0
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-factoryboy==2.1.0
pytest-randomly==3.8.0
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
zipp==3.5.0
